% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{significance_funnel}
\alias{significance_funnel}
\title{Make significance funnel plot}
\usage{
significance_funnel(
  yi,
  vi,
  sei,
  xmin = min(yi),
  xmax = max(yi),
  ymin = 0,
  ymax = max(sqrt(vi)),
  xlab = "Point estimate",
  ylab = "Estimated standard error",
  favor_positive = NA,
  est_all = NA,
  est_worst = NA,
  alpha_select = 0.05,
  plot_pooled = TRUE
)
}
\arguments{
\item{yi}{A vector of point estimates to be meta-analyzed.}

\item{vi}{A vector of estimated variances (i.e., squared standard errors) for
the point estimates.}

\item{sei}{A vector of estimated standard errors for the point estimates.
(Only one of \code{vi} or \code{sei} needs to be specified).}

\item{xmin}{x-axis (point estimate) lower limit for plot.}

\item{xmax}{x-axis (point estimate) upper limit for plot.}

\item{ymin}{y-axis (standard error) lower limit for plot.}

\item{ymax}{y-axis (standard error) upper limit for plot.}

\item{xlab}{Label for x-axis (point estimate).}

\item{ylab}{Label for y-axis (standard error).}

\item{favor_positive}{\code{TRUE} if publication bias are
assumed to favor significant positive estimates; \code{FALSE} if assumed to
favor significant negative estimates.}

\item{est_all}{Regular meta-analytic estimate among all studies (optional).}

\item{est_worst}{Worst-case meta-analytic estimate among only nonaffirmative
studies (optional).}

\item{alpha_select}{Alpha level at which an estimate's probability of being
favored by publication bias is assumed to change (i.e.,
the threshold at which study investigators, journal editors, etc., consider
an estimate to be significant).}

\item{plot_pooled}{Should the pooled estimates within all studies and within
only the nonaffirmative studies be plotted as well?}
}
\description{
Creates a modified funnel plot that distinguishes between affirmative and
nonaffirmative studies, helping to detect the extent to which the
nonaffirmative studies' point estimates are systematically smaller than the
entire set of point estimates. The estimate among only nonaffirmative studies
(gray diamond) represents a corrected estimate under worst-case publication
bias. If the gray diamond represents a negligible effect size or if it is
much smaller than the pooled estimate among all studies (black diamond), this
suggests that the meta-analysis may not be robust to extreme publication
bias. Numerical sensitivity analyses (via \code{\link[=pubbias_svalue]{pubbias_svalue()}}) should still be
carried out for more precise quantitative conclusions.
}
\details{
By default (\code{plot_pooled = TRUE}), also plots the pooled point
estimate within all studies, supplied by the user as \code{est_all} (black
diamond), and within only the nonaffirmative studies, supplied by the user
as \code{est_worst} (grey diamond). The user can calculate \code{est_all} and
\code{est_worst} using their choice of meta-analysis model. If instead these
are not supplied but \code{plot_pooled = TRUE}, these pooled estimates will
be automatically calculated using a fixed-effects (a.k.a. "common-effect")
model.
}
\examples{
##### Make Significance Funnel with User-Specified Pooled Estimates #####
# compute meta-analytic effect sizes for an example dataset
require(metafor)
dat <- metafor::escalc(measure = "RR", ai = tpos, bi = tneg, ci = cpos,
                       di = cneg, data = dat.bcg)

# favor_positive = FALSE since we think publication bias operates in favor of negative
significance_funnel(yi = dat$yi, vi = dat$vi, favor_positive = FALSE)

# by default, the meta-analytic estimates plotted are from metafor::rma.uni()
# you can instead supply estimates from another meta-analysis model
# for example, using the robust independent specification since the
# point estimates seem to be from unique papers
# require(robumeta)
# meta_all <- robumeta::robu(yi ~ 1,
#                            studynum = 1:nrow(dat),
#                            data = dat,
#                            var.eff.size = vi,
#                            small = TRUE)
#
# # worst-case meta-analysis (non-affirmative studies)
# dat$pval <- 2 * (1 - pnorm(abs(dat$yi / sqrt(dat$vi))))  # two-tailed p-value
# dat$affirm <- (dat$yi > 0) & (dat$pval < 0.05)  # is study affirmative?
# meta_worst <- robumeta::robu(yi ~ 1,
#                              studynum = 1:nrow( dat[ dat$affirm == TRUE, ] ),
#                              data = dat[ dat$affirm == TRUE, ],
#                              var.eff.size = vi,
#                              small = TRUE)
#
# meta_all <- pubbias_meta(yi = dat$yi, vi = dat$vi, selection_ratio = 1) # etc
# meta_all$fits$robust$b.r
# meta_all$data
# pubbias_svalue(yi = dat$yi, vi = dat$vi, favor_positive = FALSE, return_worst_meta = TRUE)
#
# significance_funnel(yi = dat$yi, vi = dat$vi, est_all = meta_all$b.r,
#                     est_worse = meta_worst$b.r, favor_positive = FALSE)
}
\references{
\insertRef{mathur2020}{metabias}
}
